Tree Algorithms
Here in this problem one tc is not passing, i hv tried to optimize it to the fullest but then also it is not accepted therefore i had to do little tricks then it got accepted. I dont like doing these but https://codeforces.com/profile/pllk this pllk guy the creator of cses is not willing to increase java time limit.